Refactor hermes_data organization
