+++ crates/hermes-data/src/command.rs
@@ -1,0 +1,16 @@
+// use std::sync::Arc;
+// use crate::{Item, Type};
+//
+// #[derive(Clone, Debug)]
+// pub struct Argument {
+//     pub head: Item,
+//     pub type_: Arc<Type>,
+// }
+//
+// #[derive(Clone, Debug)]
+// pub struct Command {
+//     pub head: Item,
+//     pub abstract_: bool,
+//     pub args: Vec<Argument>,
+//
+// }
