+++ crates/hermes-data/src/de/context.rs
@@ -1,9 +1,9 @@
-use crate::bit_vec::BitVec;
+use crate::de::{BitVec, Entry, Packet, ParameterValue, SequenceContainer, SequenceContainerType};
 use crate::{
-    ByteOrder, Entry, Error, MissionDatabase, Packet, Parameter, ParameterInstanceRef,
-    ParameterRef, ParameterValue, Result, SequenceContainer, SequenceContainerType, Value,
-    VariableSize,
+    ByteOrder, Error, MissionDatabase, Parameter, ParameterInstanceRef, ParameterRef, Result,
+    Value, VariableSize,
 };
+
 use std::collections::HashMap;
 use std::sync::Arc;
 use std::time::Instant;
