+++ crates/hermes-data/src/lib.rs
@@ -1,23 +1,18 @@
-mod bit_vec;
 mod calibrator;
-mod container;
-mod deserialize;
+mod command;
+pub mod de;
 mod error;
 mod framing;
-mod parameter;
 mod types;
 mod util;
 mod xtce;
+mod item;
 
 pub use calibrator::*;
-pub use container::*;
-pub use deserialize::*;
 pub use error::{Error, Result};
 pub use framing::*;
-pub use parameter::*;
 pub use types::*;
-
-use xtce::*;
+pub use item::*;
 
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -31,10 +26,10 @@ pub struct MissionDatabase {
     // command_arguments: HashMap<String, Arc<Argument>>,
     // command_containers: HashMap<String, Arc<SequenceContainer>>,
     // commands: HashMap<String, Arc<MetaCommandType>>,
-    telemetry_root: Arc<SequenceContainerType>,
+    telemetry_root: Arc<de::SequenceContainerType>,
     telemetry_parameter_types: HashMap<String, Arc<Type>>,
     telemetry_parameters: HashMap<String, Arc<Parameter>>,
-    telemetry_containers: HashMap<String, Arc<SequenceContainerType>>,
+    telemetry_containers: HashMap<String, Arc<de::SequenceContainerType>>,
 }
 
 // Compile-time checks to ensure MissionDatabase is thread-safe
@@ -62,30 +57,30 @@ impl MissionDatabase {
 
         // Pass 1: Collect unresolved parameter types
         let mut unresolved_param_types = HashMap::new();
-        collect_parameter_types(&root_path, schema, &mut unresolved_param_types);
+        xtce::collect_parameter_types(&root_path, schema, &mut unresolved_param_types);
 
         // Pass 2: Construct simple types (int, float, string, bool, enum, time)
         // Defer Binary, Array, and Aggregate types
         let (mut telemetry_parameter_types, deferred_binary, deferred_array, deferred_aggregate) =
-            construct_parameter_types_pass1(unresolved_param_types)?;
+            xtce::construct_parameter_types_pass1(unresolved_param_types)?;
 
         // Pass 3: Construct Aggregate types (now that simple types are available)
-        construct_parameter_types_pass2_aggregates(
+        xtce::construct_parameter_types_pass2_aggregates(
             deferred_aggregate,
             &mut telemetry_parameter_types,
         )?;
 
         // Pass 4: Collect unresolved parameters
         let mut unresolved_parameters = HashMap::new();
-        collect_parameters(&root_path, schema, &mut unresolved_parameters);
+        xtce::collect_parameters(&root_path, schema, &mut unresolved_parameters);
 
         // Pass 5: Construct parameters with simple/aggregate types
         // Parameters that reference binary/array types will be deferred
         let (mut telemetry_parameters, still_unresolved_params) =
-            construct_parameters(unresolved_parameters, &telemetry_parameter_types);
+            xtce::construct_parameters(unresolved_parameters, &telemetry_parameter_types);
 
         // Pass 6: Construct Binary and Array types (now that parameters are available for resolution)
-        construct_parameter_types_pass3_binary_array(
+        xtce::construct_parameter_types_pass3_binary_array(
             deferred_binary,
             deferred_array,
             &mut telemetry_parameter_types,
@@ -93,7 +88,7 @@ impl MissionDatabase {
         )?;
 
         // Pass 7: Construct remaining parameters (those that reference binary/array types)
-        construct_remaining_parameters(
+        xtce::construct_remaining_parameters(
             still_unresolved_params,
             &telemetry_parameter_types,
             &mut telemetry_parameters,
@@ -101,10 +96,10 @@ impl MissionDatabase {
 
         // Pass 8: Collect all containers with their unresolved references
         let mut unresolved_containers = HashMap::new();
-        collect_containers(&root_path, schema, &mut unresolved_containers);
+        xtce::collect_containers(&root_path, schema, &mut unresolved_containers);
 
         // Pass 9: Build dependency graph and topological sort
-        let (sorted_names, dependencies) = build_dependency_graph(&unresolved_containers)?;
+        let (sorted_names, dependencies) = xtce::build_dependency_graph(&unresolved_containers)?;
 
         if sorted_names.is_empty() {
             return Err(Error::InvalidXtce(
@@ -115,7 +110,7 @@ impl MissionDatabase {
         let root_name = sorted_names[0].clone();
 
         // Pass 10: Construct containers in dependency order (all references fully resolved)
-        let telemetry_containers = construct_containers(
+        let telemetry_containers = xtce::construct_containers(
             unresolved_containers,
             sorted_names,
             dependencies,
@@ -132,7 +127,7 @@ impl MissionDatabase {
         })
     }
 
-    pub fn telemetry_root(&self) -> &Arc<SequenceContainerType> {
+    pub fn telemetry_root(&self) -> &Arc<de::SequenceContainerType> {
         &self.telemetry_root
     }
 
@@ -152,11 +147,11 @@ impl MissionDatabase {
         &self.telemetry_parameter_types
     }
 
-    pub fn get_telemetry_container(&self, name: &str) -> Option<&Arc<SequenceContainerType>> {
+    pub fn get_telemetry_container(&self, name: &str) -> Option<&Arc<de::SequenceContainerType>> {
         self.telemetry_containers.get(name)
     }
 
-    pub fn all_telemetry_containers(&self) -> &HashMap<String, Arc<SequenceContainerType>> {
+    pub fn all_telemetry_containers(&self) -> &HashMap<String, Arc<de::SequenceContainerType>> {
         &self.telemetry_containers
     }
 }
